Re-organize the module a bit
