+++ src/rootfilespec/bootstrap/RAnchor.py
@@ -1,14 +1,10 @@
-from __future__ import annotations
-
 from typing import Annotated
 
-from rootfilespec.bootstrap.envelopebase import (
-    REnvelopeLink,
-)
-from rootfilespec.bootstrap.envelopes import FooterEnvelope, HeaderEnvelope
-from rootfilespec.bootstrap.RLocator import LargeLocator
 from rootfilespec.bootstrap.TObject import StreamedObject
 from rootfilespec.dispatch import DICTIONARY
+from rootfilespec.rntuple import FooterEnvelope, HeaderEnvelope
+from rootfilespec.rntuple.envelope import REnvelopeLink
+from rootfilespec.rntuple.RLocator import LargeLocator
 from rootfilespec.structutil import (
     DataFetcher,
     Fmt,
