+++ src/rootfilespec/rntuple/header.py
@@ -1,0 +1,10 @@
+from rootfilespec.rntuple.envelope import ENVELOPE_TYPE_MAP, REnvelope
+from rootfilespec.structutil import serializable
+
+
+@serializable
+class HeaderEnvelope(REnvelope):
+    """A class representing the RNTuple Header Envelope payload structure"""
+
+
+ENVELOPE_TYPE_MAP[0x01] = "HeaderEnvelope"
